requirements.txt
beautifulsoup4==4.6.3
bs4==0.0.1
certifi==2018.8.24
chardet==3.0.4
idna==2.7
lxml==4.2.5
mysql-connector==2.1.6
numpy==1.15.1
pandas==0.23.4
python-dateutil==2.7.3
pytz==2018.5
requests==2.19.1
six==1.11.0
SQLAlchemy==1.2.11
tqdm==4.26.0
urllib3==1.23
